+++ ui/src/external/monaco.fluxLang.ts
@@ -1,0 +1,17 @@
+import initialize from 'src/external/monaco'
+
+initialize().then((monaco: any) => {
+  monaco.languages.register({id: 'flux'})
+
+  monaco.languages.setMonarchTokensProvider('flux', {
+    keywords: ['from', 'range', 'filter', 'to'],
+    tokenizer: {
+      root: [
+        [
+          /[a-z_$][\w$]*/,
+          {cases: {'@keywords': 'keyword', '@default': 'identifier'}},
+        ],
+      ],
+    },
+  })
+})
