+++ ui/src/external/monaco.fluxTheme.ts
@@ -1,0 +1,9 @@
+import initialize from 'src/external/monaco'
+
+initialize().then((monaco: any) => {
+  monaco.editor.defineTheme('fluxTheme', {
+    base: 'vs',
+    inherit: false,
+    rules: [{token: 'keyword', foreground: 'ff00ff', fontStyle: 'bold'}],
+  })
+})
