+++ ui/src/shared/components/FluxMonacoEditor.tsx
@@ -1,7 +1,11 @@
 // Libraries
 import React, {FC} from 'react'
+import 'src/external/monaco.fluxLang'
+import 'src/external/monaco.fluxTheme'
+import 'src/external/monaco.fluxCompletions'
+
+// Components
 import Editor from '@monaco-editor/react'
-import 'src/external/monaco.flux'
 
 interface Props {
   script: string
@@ -10,7 +14,7 @@ interface Props {
 const FluxEditorMonaco: FC<Props> = ({script}) => {
   return (
     <div className="time-machine-editor" data-testid="flux-editor">
-      <Editor height="90vh" language="Flux" value={script} />
+      <Editor height="90vh" language="flux" value={script} />
     </div>
   )
 }
